Create BTC MESSAGE FINDER
THIS UNTESTED CODE MIGHT...

This code allows you to search for Bitcoin transactions that contain a specific message in the comment field. It uses the listtransactions RPC method to retrieve a list of transaction details for a particular Bitcoin address, and then filters the list to only include transactions that contain the specified message. The filtered list of transactions is then printed to the console, along with the base64-encoded data extracted from the message.

To use this code, you will need to have a Bitcoin Core node running on your local machine and configured to accept RPC calls. You will also need to have Python 3 and the requests library installed. Once you have these dependencies installed, you can simply copy and paste the code into a Python script and run it from the command line.